+++ asset_test.go
@@ -13,31 +13,6 @@ import (
 	"github.com/limrun-inc/go-sdk/option"
 )
 
-func TestAssetNew(t *testing.T) {
-	t.Skip("Prism tests are disabled")
-	baseURL := "http://localhost:4010"
-	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
-		baseURL = envURL
-	}
-	if !testutil.CheckTestServer(t, baseURL) {
-		return
-	}
-	client := limrun.NewClient(
-		option.WithBaseURL(baseURL),
-		option.WithAPIKey("My API Key"),
-	)
-	_, err := client.Assets.New(context.TODO(), limrun.AssetNewParams{
-		Name: "name",
-	})
-	if err != nil {
-		var apierr *limrun.Error
-		if errors.As(err, &apierr) {
-			t.Log(string(apierr.DumpRequest(true)))
-		}
-		t.Fatalf("err should be nil: %s", err.Error())
-	}
-}
-
 func TestAssetGetWithOptionalParams(t *testing.T) {
 	t.Skip("Prism tests are disabled")
 	baseURL := "http://localhost:4010"
@@ -86,6 +61,31 @@ func TestAssetListWithOptionalParams(t *testing.T) {
 		IncludeUploadURL:   limrun.Bool(true),
 		Md5Filter:          limrun.String("md5Filter"),
 		NameFilter:         limrun.String("nameFilter"),
+	})
+	if err != nil {
+		var apierr *limrun.Error
+		if errors.As(err, &apierr) {
+			t.Log(string(apierr.DumpRequest(true)))
+		}
+		t.Fatalf("err should be nil: %s", err.Error())
+	}
+}
+
+func TestAssetGetOrNew(t *testing.T) {
+	t.Skip("Prism tests are disabled")
+	baseURL := "http://localhost:4010"
+	if envURL, ok := os.LookupEnv("TEST_API_BASE_URL"); ok {
+		baseURL = envURL
+	}
+	if !testutil.CheckTestServer(t, baseURL) {
+		return
+	}
+	client := limrun.NewClient(
+		option.WithBaseURL(baseURL),
+		option.WithAPIKey("My API Key"),
+	)
+	_, err := client.Assets.GetOrNew(context.TODO(), limrun.AssetGetOrNewParams{
+		Name: "name",
 	})
 	if err != nil {
 		var apierr *limrun.Error