+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## 0.2.0 (2025-09-08)
+
+Full Changelog: [v0.1.0...v0.2.0](https://github.com/limrun-inc/go-sdk/compare/v0.1.0...v0.2.0)
+
+### Features
+
+* **assets:** add getOrUpload method and example ([1debc0c](https://github.com/limrun-inc/go-sdk/commit/1debc0cae08deb62f7e9a5335f7b182d94b668b8))
+* **assets:** rename create to getOrCreate ([2df7ff5](https://github.com/limrun-inc/go-sdk/commit/2df7ff5817290e496f848bcc8ad0c0dd1a8732b1))
+* **examples:** add server example ([0d5b692](https://github.com/limrun-inc/go-sdk/commit/0d5b692334b3b3da0bc4351d4ccbc7c3a0584946))
+* **tunnel:** add tunnel package and example ([4e84071](https://github.com/limrun-inc/go-sdk/commit/4e840719bdc54e0d15d96e7ba0857f2cc2b8f5a4))
+
 ## 0.1.0 (2025-09-08)
 
 Full Changelog: [v0.0.1...v0.1.0](https://github.com/limrun-inc/go-sdk/compare/v0.0.1...v0.1.0)